ingestion_job/pyproject.toml
[project]
name = "ingestion-job"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "lxml>=6.0.2",
    "neo4j>=6.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
